Move diff code into web_monitoring/diff
