+++ web_monitoring/diff/links_diff.py
@@ -1,7 +1,7 @@
 import html5_parser
 from .content_type import raise_if_not_diffable_html
 from .differs import compute_dmp_diff
-from web_monitoring.utils import get_color_palette
+from ..utils import get_color_palette
 from difflib import SequenceMatcher
 from .html_diff_render import (get_title, _html_for_dmp_operation,
                                undiffable_content_tags)
